requirements.txt
numpy>=2.1.0
PyQt5>=5.15.11
pyqtgraph>=0.13.7
pyserial>=3.5